refactor textdet postprocess using build_postprocessor
