+++ mmocr/models/textdet/dense_heads/__init__.py
@@ -1,13 +1,13 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .base_head import BaseHead
 from .db_head import DBHead
 from .drrg_head import DRRGHead
 from .fce_head import FCEHead
-from .head_mixin import HeadMixin
 from .pan_head import PANHead
 from .pse_head import PSEHead
 from .textsnake_head import TextSnakeHead
 
 __all__ = [
-    'PSEHead', 'PANHead', 'DBHead', 'FCEHead', 'HeadMixin', 'TextSnakeHead',
-    'DRRGHead'
+    'PSEHead', 'PANHead', 'DBHead', 'FCEHead', 'TextSnakeHead', 'DRRGHead',
+    'BaseHead'
 ]
