+++ mmocr/models/textdet/dense_heads/pse_head.py
@@ -10,8 +10,6 @@ class PSEHead(PANHead):
     Args:
         in_channels (list[int]): A list of 4 numbers of input channels.
         out_channels (int): Number of output channels.
-        text_repr_type (str): Use polygon or quad to represent. Available
-            options are "poly" or "quad".
         downsample_ratio (float): Downsample ratio.
         loss (dict): Configuration dictionary for loss type. Supported loss
             types are "PANLoss" and "PSELoss".
@@ -19,22 +17,25 @@ class PSEHead(PANHead):
         init_cfg (dict or list[dict], optional): Initialization configs.
     """
 
-    def __init__(
-            self,
-            in_channels,
-            out_channels,
-            text_repr_type='poly',  # 'poly' or 'quad'
-            downsample_ratio=0.25,
-            loss=dict(type='PSELoss'),
-            train_cfg=None,
-            test_cfg=None,
-            init_cfg=None):
+    def __init__(self,
+                 in_channels,
+                 out_channels,
+                 downsample_ratio=0.25,
+                 loss=dict(type='PSELoss'),
+                 postprocessor=dict(
+                     type='PSEPostprocessor', text_repr_type='poly'),
+                 train_cfg=None,
+                 test_cfg=None,
+                 init_cfg=None,
+                 **kwargs):
+
         super().__init__(
             in_channels=in_channels,
             out_channels=out_channels,
-            text_repr_type=text_repr_type,
             downsample_ratio=downsample_ratio,
             loss=loss,
+            postprocessor=postprocessor,
             train_cfg=train_cfg,
             test_cfg=test_cfg,
-            init_cfg=init_cfg)
+            init_cfg=init_cfg,
+            **kwargs)
