+++ mmocr/models/textdet/dense_heads/textsnake_head.py
@@ -2,22 +2,20 @@
 import torch.nn as nn
 from mmcv.runner import BaseModule
 
-from mmocr.models.builder import HEADS, build_loss
-from . import HeadMixin
+from mmocr.models.builder import HEADS
+from .base_head import BaseHead
 
 
 @HEADS.register_module()
-class TextSnakeHead(HeadMixin, BaseModule):
-    """The class for TextSnake head.
+class TextSnakeHead(BaseHead, BaseModule):
+    """The class for TextSnake head: TextSnake: A Flexible Representation for
+    Detecting Text of Arbitrary Shapes.
 
     TextSnake: `A Flexible Representation for Detecting Text of Arbitrary
     Shapes <https://arxiv.org/abs/1807.01544>`_.
 
     Args:
         in_channels (int): Number of input channels.
-        decoding_type (str): Decoding type. It usually should not be changed.
-        text_repr_type (str): Use polygon or quad to represent. Available
-            options are "poly" or "quad".
         loss (dict): Configuration dictionary for loss type.
         train_cfg, test_cfg: Depreciated.
         init_cfg (dict or list[dict], optional): Initialization configs.
@@ -25,9 +23,11 @@ class TextSnakeHead(HeadMixin, BaseModule):
 
     def __init__(self,
                  in_channels,
-                 decoding_type='textsnake',
-                 text_repr_type='poly',
+                 out_channels=5,
+                 downsample_ratio=1.0,
                  loss=dict(type='TextSnakeLoss'),
+                 postprocessor=dict(
+                     type='TextSnakePostprocessor', text_repr_type='poly'),
                  train_cfg=None,
                  test_cfg=None,
                  init_cfg=dict(
@@ -35,15 +35,13 @@ class TextSnakeHead(HeadMixin, BaseModule):
                      override=dict(name='out_conv'),
                      mean=0,
                      std=0.01)):
-        super().__init__(init_cfg=init_cfg)
+        BaseModule.__init__(self, init_cfg=init_cfg)
+        BaseHead.__init__(self, loss, postprocessor)
 
         assert isinstance(in_channels, int)
         self.in_channels = in_channels
-        self.out_channels = 5
-        self.downsample_ratio = 1.0
-        self.decoding_type = decoding_type
-        self.text_repr_type = text_repr_type
-        self.loss_module = build_loss(loss)
+        self.out_channels = out_channels
+        self.downsample_ratio = downsample_ratio
         self.train_cfg = train_cfg
         self.test_cfg = test_cfg
