+++ mmocr/models/textdet/detectors/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .base_detector import BaseDetector
 from .dbnet import DBNet
 from .drrg import DRRG
 from .fcenet import FCENet
@@ -6,10 +7,9 @@ from .ocr_mask_rcnn import OCRMaskRCNN
 from .panet import PANet
 from .psenet import PSENet
 from .single_stage_text_detector import SingleStageTextDetector
-from .text_detector_mixin import TextDetectorMixin
 from .textsnake import TextSnake
 
 __all__ = [
-    'TextDetectorMixin', 'SingleStageTextDetector', 'OCRMaskRCNN', 'DBNet',
-    'PANet', 'PSENet', 'TextSnake', 'FCENet', 'DRRG'
+    'BaseDetector', 'SingleStageTextDetector', 'OCRMaskRCNN', 'DBNet', 'PANet',
+    'PSENet', 'TextSnake', 'FCENet', 'DRRG'
 ]
