+++ mmocr/models/textdet/detectors/drrg.py
@@ -1,13 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from mmocr.models.builder import DETECTORS
-from mmocr.models.textdet.detectors.single_stage_text_detector import \
-    SingleStageTextDetector
-from mmocr.models.textdet.detectors.text_detector_mixin import \
-    TextDetectorMixin
+from .base_detector import BaseDetector
+from .single_stage_text_detector import SingleStageTextDetector
 
 
 @DETECTORS.register_module()
-class DRRG(TextDetectorMixin, SingleStageTextDetector):
+class DRRG(BaseDetector, SingleStageTextDetector):
     """The class for implementing DRRG text detector. Deep Relational Reasoning
     Graph Network for Arbitrary Shape Text Detection.
 
@@ -26,7 +24,7 @@ class DRRG(TextDetectorMixin, SingleStageTextDetector):
         SingleStageTextDetector.__init__(self, backbone, neck, bbox_head,
                                          train_cfg, test_cfg, pretrained,
                                          init_cfg)
-        TextDetectorMixin.__init__(self, show_score)
+        BaseDetector.__init__(self, show_score)
 
     def forward_train(self, img, img_metas, **kwargs):
         """
