+++ mmocr/models/textdet/detectors/psenet.py
@@ -1,13 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from mmocr.models.builder import DETECTORS
-from mmocr.models.textdet.detectors.single_stage_text_detector import \
-    SingleStageTextDetector
-from mmocr.models.textdet.detectors.text_detector_mixin import \
-    TextDetectorMixin
+from .base_detector import BaseDetector
+from .single_stage_text_detector import SingleStageTextDetector
 
 
 @DETECTORS.register_module()
-class PSENet(TextDetectorMixin, SingleStageTextDetector):
+class PSENet(BaseDetector, SingleStageTextDetector):
     """The class for implementing PSENet text detector: Shape Robust Text
     Detection with Progressive Scale Expansion Network.
 
@@ -26,4 +24,4 @@ class PSENet(TextDetectorMixin, SingleStageTextDetector):
         SingleStageTextDetector.__init__(self, backbone, neck, bbox_head,
                                          train_cfg, test_cfg, pretrained,
                                          init_cfg)
-        TextDetectorMixin.__init__(self, show_score)
+        BaseDetector.__init__(self, show_score)
