+++ mmocr/models/textdet/postprocess/__init__.py
@@ -1,4 +1,14 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .wrapper import decode
+from .base_postprocessor import BasePostprocessor
+from .db_postprocessor import DBPostprocessor
+from .drrg_postprocessor import DrrgPostprocessor
+from .fce_postprocessor import FCEPostprocessor
+from .pan_postprocessor import PANPostprocessor
+from .pse_postprocessor import PSEPostprocessor
+from .textsnake_postprocessor import TextSnakePostprocessor
 
-__all__ = ['decode']
+__all__ = [
+    'BasePostprocessor', 'PSEPostprocessor', 'PANPostprocessor',
+    'DBPostprocessor', 'DrrgPostprocessor', 'FCEPostprocessor',
+    'TextSnakePostprocessor'
+]
