+++ mmocr/models/textdet/postprocess/wrapper.py
@@ -5,150 +5,15 @@ import operator
 import cv2
 import numpy as np
 import pyclipper
-import torch
-from mmcv.ops import contour_expand, pixel_group
 from numpy.fft import ifft
 from numpy.linalg import norm
 from shapely.geometry import Polygon
-from skimage.morphology import skeletonize
 
-from mmocr.core import points2boundary
 from mmocr.core.evaluation.utils import boundary_iou
 
 
 def filter_instance(area, confidence, min_area, min_confidence):
     return bool(area < min_area or confidence < min_confidence)
-
-
-def decode(
-        decoding_type='pan',  # 'pan' or 'pse'
-        **kwargs):
-    if decoding_type == 'pan':
-        return pan_decode(**kwargs)
-    if decoding_type == 'pse':
-        return pse_decode(**kwargs)
-    if decoding_type == 'db':
-        return db_decode(**kwargs)
-    if decoding_type == 'textsnake':
-        return textsnake_decode(**kwargs)
-    if decoding_type == 'fcenet':
-        return fcenet_decode(**kwargs)
-    if decoding_type == 'drrg':
-        return drrg_decode(**kwargs)
-
-    raise NotImplementedError
-
-
-def pan_decode(preds,
-               text_repr_type='poly',
-               min_text_confidence=0.5,
-               min_kernel_confidence=0.5,
-               min_text_avg_confidence=0.85,
-               min_text_area=16):
-    """Convert scores to quadrangles via post processing in PANet. This is
-    partially adapted from https://github.com/WenmuZhou/PAN.pytorch.
-
-    Args:
-        preds (tensor): The head output tensor of size 6xHxW.
-        text_repr_type (str): The boundary encoding type 'poly' or 'quad'.
-        min_text_confidence (float): The minimal text confidence.
-        min_kernel_confidence (float): The minimal kernel confidence.
-        min_text_avg_confidence (float): The minimal text average confidence.
-        min_text_area (int): The minimal text instance region area.
-    Returns:
-        boundaries: (list[list[float]]): The instance boundary and its
-            instance confidence list.
-    """
-    preds[:2, :, :] = torch.sigmoid(preds[:2, :, :])
-    preds = preds.detach().cpu().numpy()
-
-    text_score = preds[0].astype(np.float32)
-    text = preds[0] > min_text_confidence
-    kernel = (preds[1] > min_kernel_confidence) * text
-    embeddings = preds[2:].transpose((1, 2, 0))  # (h, w, 4)
-
-    region_num, labels = cv2.connectedComponents(
-        kernel.astype(np.uint8), connectivity=4)
-    contours, _ = cv2.findContours((kernel * 255).astype(np.uint8),
-                                   cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)
-    kernel_contours = np.zeros(text.shape, dtype='uint8')
-    cv2.drawContours(kernel_contours, contours, -1, 255)
-    text_points = pixel_group(text_score, text, embeddings, labels,
-                              kernel_contours, region_num,
-                              min_text_avg_confidence)
-
-    boundaries = []
-    for text_inx, text_point in enumerate(text_points):
-        text_confidence = text_point[0]
-        text_point = text_point[2:]
-        text_point = np.array(text_point, dtype=int).reshape(-1, 2)
-        area = text_point.shape[0]
-
-        if filter_instance(area, text_confidence, min_text_area,
-                           min_text_avg_confidence):
-            continue
-        vertices_confidence = points2boundary(text_point, text_repr_type,
-                                              text_confidence)
-        if vertices_confidence is not None:
-            boundaries.append(vertices_confidence)
-
-    return boundaries
-
-
-def pse_decode(preds,
-               text_repr_type='poly',
-               min_kernel_confidence=0.5,
-               min_text_avg_confidence=0.85,
-               min_kernel_area=0,
-               min_text_area=16):
-    """Decoding predictions of PSENet to instances. This is partially adapted
-    from https://github.com/whai362/PSENet.
-
-    Args:
-        preds (tensor): The head output tensor of size nxHxW.
-        text_repr_type (str): The boundary encoding type 'poly' or 'quad'.
-        min_text_confidence (float): The minimal text confidence.
-        min_kernel_confidence (float): The minimal kernel confidence.
-        min_text_avg_confidence (float): The minimal text average confidence.
-        min_kernel_area (int): The minimal text kernel area.
-        min_text_area (int): The minimal text instance region area.
-    Returns:
-        boundaries: (list[list[float]]): The instance boundary and its
-            instance confidence list.
-    """
-    preds = torch.sigmoid(preds)  # text confidence
-
-    score = preds[0, :, :]
-    masks = preds > min_kernel_confidence
-    text_mask = masks[0, :, :]
-    kernel_masks = masks[0:, :, :] * text_mask
-
-    score = score.data.cpu().numpy().astype(np.float32)  # to numpy
-
-    kernel_masks = kernel_masks.data.cpu().numpy().astype(np.uint8)  # to numpy
-
-    region_num, labels = cv2.connectedComponents(
-        kernel_masks[-1], connectivity=4)
-
-    # labels = pse(kernel_masks, min_kernel_area)
-    labels = contour_expand(kernel_masks, labels, min_kernel_area, region_num)
-    labels = np.array(labels)
-    label_num = np.max(labels)
-    boundaries = []
-    for i in range(1, label_num + 1):
-        points = np.array(np.where(labels == i)).transpose((1, 0))[:, ::-1]
-        area = points.shape[0]
-        score_instance = np.mean(score[labels == i])
-        if filter_instance(area, score_instance, min_text_area,
-                           min_text_avg_confidence):
-            continue
-
-        vertices_confidence = points2boundary(points, text_repr_type,
-                                              score_instance)
-        if vertices_confidence is not None:
-            boundaries.append(vertices_confidence)
-
-    return boundaries
 
 
 def box_score_fast(bitmap, _box):
@@ -173,72 +38,6 @@ def unclip(box, unclip_ratio=1.5):
     offset.AddPath(box, pyclipper.JT_ROUND, pyclipper.ET_CLOSEDPOLYGON)
     expanded = np.array(offset.Execute(distance))
     return expanded
-
-
-def db_decode(preds,
-              text_repr_type='poly',
-              mask_thr=0.3,
-              min_text_score=0.3,
-              min_text_width=5,
-              unclip_ratio=1.5,
-              max_candidates=3000):
-    """Decoding predictions of DbNet to instances. This is partially adapted
-    from https://github.com/MhLiao/DB.
-
-    Args:
-        preds (Tensor): The head output tensor of size nxHxW.
-        text_repr_type (str): The boundary encoding type 'poly' or 'quad'.
-        mask_thr (float): The mask threshold value for binarization.
-        min_text_score (float): The threshold value for converting binary map
-            to shrink text regions.
-        min_text_width (int): The minimum width of boundary polygon/box
-            predicted.
-        unclip_ratio (float): The unclip ratio for text regions dilation.
-        max_candidates (int): The maximum candidate number.
-
-    Returns:
-        boundaries: (list[list[float]]): The predicted text boundaries.
-    """
-    prob_map = preds[0, :, :]
-    text_mask = prob_map > mask_thr
-
-    score_map = prob_map.data.cpu().numpy().astype(np.float32)
-    text_mask = text_mask.data.cpu().numpy().astype(np.uint8)  # to numpy
-
-    contours, _ = cv2.findContours((text_mask * 255).astype(np.uint8),
-                                   cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
-
-    boundaries = []
-    for i, poly in enumerate(contours):
-        if i > max_candidates:
-            break
-        epsilon = 0.01 * cv2.arcLength(poly, True)
-        approx = cv2.approxPolyDP(poly, epsilon, True)
-        points = approx.reshape((-1, 2))
-        if points.shape[0] < 4:
-            continue
-        score = box_score_fast(score_map, points)
-        if score < min_text_score:
-            continue
-        poly = unclip(points, unclip_ratio=unclip_ratio)
-        if len(poly) == 0 or isinstance(poly[0], list):
-            continue
-        poly = poly.reshape(-1, 2)
-
-        if text_repr_type == 'quad':
-            poly = points2boundary(poly, text_repr_type, score, min_text_width)
-        elif text_repr_type == 'poly':
-            poly = poly.flatten().tolist()
-            if score is not None:
-                poly = poly + [score]
-            if len(poly) < 8:
-                poly = None
-        else:
-            raise ValueError(f'Invalid text repr type {text_repr_type}')
-
-        if poly is not None:
-            boundaries.append(poly)
-    return boundaries
 
 
 def fill_hole(input_mask):
@@ -313,177 +112,6 @@ def merge_disks(disks, disk_overlap_thr):
     merged_disks = np.vstack(merged_disks)
 
     return merged_disks
-
-
-def textsnake_decode(preds,
-                     text_repr_type='poly',
-                     min_text_region_confidence=0.6,
-                     min_center_region_confidence=0.2,
-                     min_center_area=30,
-                     disk_overlap_thr=0.03,
-                     radius_shrink_ratio=1.03):
-    """Decoding predictions of TextSnake to instances. This was partially
-    adapted from https://github.com/princewang1994/TextSnake.pytorch.
-
-    Args:
-        preds (tensor): The head output tensor of size 6xHxW.
-        text_repr_type (str): The boundary encoding type 'poly' or 'quad'.
-        min_text_region_confidence (float): The confidence threshold of text
-            region in TextSnake.
-        min_center_region_confidence (float): The confidence threshold of text
-            center region in TextSnake.
-        min_center_area (int): The minimal text center region area.
-        disk_overlap_thr (float): The radius overlap threshold for merging
-            disks.
-        radius_shrink_ratio (float): The shrink ratio of ordered disks radii.
-
-    Returns:
-        boundaries (list[list[float]]): The instance boundary and its
-            instance confidence list.
-    """
-    assert text_repr_type == 'poly'
-    preds[:2, :, :] = torch.sigmoid(preds[:2, :, :])
-    preds = preds.detach().cpu().numpy()
-
-    pred_text_score = preds[0]
-    pred_text_mask = pred_text_score > min_text_region_confidence
-    pred_center_score = preds[1] * pred_text_score
-    pred_center_mask = pred_center_score > min_center_region_confidence
-    pred_sin = preds[2]
-    pred_cos = preds[3]
-    pred_radius = preds[4]
-    mask_sz = pred_text_mask.shape
-
-    scale = np.sqrt(1.0 / (pred_sin**2 + pred_cos**2 + 1e-8))
-    pred_sin = pred_sin * scale
-    pred_cos = pred_cos * scale
-
-    pred_center_mask = fill_hole(pred_center_mask).astype(np.uint8)
-    center_contours, _ = cv2.findContours(pred_center_mask, cv2.RETR_TREE,
-                                          cv2.CHAIN_APPROX_SIMPLE)
-
-    boundaries = []
-    for contour in center_contours:
-        if cv2.contourArea(contour) < min_center_area:
-            continue
-        instance_center_mask = np.zeros(mask_sz, dtype=np.uint8)
-        cv2.drawContours(instance_center_mask, [contour], -1, 1, -1)
-        skeleton = skeletonize(instance_center_mask)
-        skeleton_yx = np.argwhere(skeleton > 0)
-        y, x = skeleton_yx[:, 0], skeleton_yx[:, 1]
-        cos = pred_cos[y, x].reshape((-1, 1))
-        sin = pred_sin[y, x].reshape((-1, 1))
-        radius = pred_radius[y, x].reshape((-1, 1))
-
-        center_line_yx = centralize(skeleton_yx, cos, -sin, radius,
-                                    instance_center_mask)
-        y, x = center_line_yx[:, 0], center_line_yx[:, 1]
-        radius = (pred_radius[y, x] * radius_shrink_ratio).reshape((-1, 1))
-        score = pred_center_score[y, x].reshape((-1, 1))
-        instance_disks = np.hstack([np.fliplr(center_line_yx), radius, score])
-        instance_disks = merge_disks(instance_disks, disk_overlap_thr)
-
-        instance_mask = np.zeros(mask_sz, dtype=np.uint8)
-        for x, y, radius, score in instance_disks:
-            if radius > 1:
-                cv2.circle(instance_mask, (int(x), int(y)), int(radius), 1, -1)
-        contours, _ = cv2.findContours(instance_mask, cv2.RETR_TREE,
-                                       cv2.CHAIN_APPROX_SIMPLE)
-
-        score = np.sum(instance_mask * pred_text_score) / (
-            np.sum(instance_mask) + 1e-8)
-        if (len(contours) > 0 and cv2.contourArea(contours[0]) > 0
-                and contours[0].size > 8):
-            boundary = contours[0].flatten().tolist()
-            boundaries.append(boundary + [score])
-
-    return boundaries
-
-
-def fcenet_decode(preds,
-                  fourier_degree,
-                  num_reconstr_points,
-                  scale,
-                  alpha=1.0,
-                  beta=2.0,
-                  text_repr_type='poly',
-                  score_thr=0.3,
-                  nms_thr=0.1):
-    """Decoding predictions of FCENet to instances.
-
-    Args:
-        preds (list(Tensor)): The head output tensors.
-        fourier_degree (int): The maximum Fourier transform degree k.
-        num_reconstr_points (int): The points number of the polygon
-            reconstructed from predicted Fourier coefficients.
-        scale (int): The down-sample scale of the prediction.
-        alpha (float) : The parameter to calculate final scores. Score_{final}
-                = (Score_{text region} ^ alpha)
-                * (Score_{text center region}^ beta)
-        beta (float) : The parameter to calculate final score.
-        text_repr_type (str):  Boundary encoding type 'poly' or 'quad'.
-        score_thr (float) : The threshold used to filter out the final
-            candidates.
-        nms_thr (float) :  The threshold of nms.
-
-    Returns:
-        boundaries (list[list[float]]): The instance boundary and confidence
-            list.
-    """
-    assert isinstance(preds, list)
-    assert len(preds) == 2
-    assert text_repr_type in ['poly', 'quad']
-
-    cls_pred = preds[0][0]
-    tr_pred = cls_pred[0:2].softmax(dim=0).data.cpu().numpy()
-    tcl_pred = cls_pred[2:].softmax(dim=0).data.cpu().numpy()
-
-    reg_pred = preds[1][0].permute(1, 2, 0).data.cpu().numpy()
-    x_pred = reg_pred[:, :, :2 * fourier_degree + 1]
-    y_pred = reg_pred[:, :, 2 * fourier_degree + 1:]
-
-    score_pred = (tr_pred[1]**alpha) * (tcl_pred[1]**beta)
-    tr_pred_mask = (score_pred) > score_thr
-    tr_mask = fill_hole(tr_pred_mask)
-
-    tr_contours, _ = cv2.findContours(
-        tr_mask.astype(np.uint8), cv2.RETR_TREE,
-        cv2.CHAIN_APPROX_SIMPLE)  # opencv4
-
-    mask = np.zeros_like(tr_mask)
-    boundaries = []
-    for cont in tr_contours:
-        deal_map = mask.copy().astype(np.int8)
-        cv2.drawContours(deal_map, [cont], -1, 1, -1)
-
-        score_map = score_pred * deal_map
-        score_mask = score_map > 0
-        xy_text = np.argwhere(score_mask)
-        dxy = xy_text[:, 1] + xy_text[:, 0] * 1j
-
-        x, y = x_pred[score_mask], y_pred[score_mask]
-        c = x + y * 1j
-        c[:, fourier_degree] = c[:, fourier_degree] + dxy
-        c *= scale
-
-        polygons = fourier2poly(c, num_reconstr_points)
-        score = score_map[score_mask].reshape(-1, 1)
-        polygons = poly_nms(np.hstack((polygons, score)).tolist(), nms_thr)
-
-        boundaries = boundaries + polygons
-
-    boundaries = poly_nms(boundaries, nms_thr)
-
-    if text_repr_type == 'quad':
-        new_boundaries = []
-        for boundary in boundaries:
-            poly = np.array(boundary[:-1]).reshape(-1, 2).astype(np.float32)
-            score = boundary[-1]
-            points = cv2.boxPoints(cv2.minAreaRect(poly))
-            points = np.int0(points)
-            new_boundaries.append(points.reshape(-1).tolist() + [score])
-
-    return boundaries
 
 
 def poly_nms(polygons, threshold):
@@ -850,32 +478,5 @@ def comps2boundaries(text_comps, comp_pred_labels):
 
         boundary = [p for coord in boundary_points for p in coord] + [score]
         boundaries.append(boundary)
-
-    return boundaries
-
-
-def drrg_decode(edges, scores, text_comps, link_thr):
-    """Merge text components and construct boundaries of text instances.
-
-    Args:
-        edges (ndarray): The edge array of shape N * 2, each row is a node
-            index pair that makes up an edge in graph.
-        scores (ndarray): The edge score array.
-        text_comps (ndarray): The text components.
-        link_thr (float): The edge score threshold.
-
-    Returns:
-        boundaries (List[list[float]]): The predicted boundaries of text
-            instances.
-    """
-    assert len(edges) == len(scores)
-    assert text_comps.ndim == 2
-    assert text_comps.shape[1] == 9
-    assert isinstance(link_thr, float)
-    vertices, score_dict = graph_propagation(edges, scores, text_comps)
-    clusters = connected_components(vertices, score_dict, link_thr)
-    pred_labels = clusters2labels(clusters, text_comps.shape[0])
-    text_comps, pred_labels = remove_single(text_comps, pred_labels)
-    boundaries = comps2boundaries(text_comps, pred_labels)
 
     return boundaries
